Clean up the micro bench mark
