+++ benchmark/xetla_kernel/bgemm/CMakeLists.txt
@@ -1,12 +1,1 @@
-set(TARGET bgemm_performance)
-
-#The SYCL kernel should be compiled with "-fsycl"
-set_source_files_properties(bgemm.cpp PROPERTIES COMPILE_FLAGS "-fsycl")
-
-add_library(${TARGET} STATIC bgemm.cpp)
-target_compile_options(${TARGET} PUBLIC "-fPIC")
-target_link_libraries(${TARGET} PUBLIC "${TORCH_LIBRARIES}")
-target_link_libraries(${TARGET} PUBLIC "${TORCH_IPEX_LIBRARIES}")
-target_include_directories(${TARGET} PUBLIC "${Python_INCLUDE_DIRS}")
-target_include_directories(${TARGET} PUBLIC "${TORCH_IPEX_INCLUDE_DIRS}")
-target_include_directories(${TARGET} PUBLIC "${XeTLALibrary_INCLUDE_DIR}")
+target_include_directories(xetla_kernel PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
